requirements.txt
boto3
botocore
slack-sdk
PyGithub
opensearch-py
langchain
langchain-community
requests
urllib3
numpy
pandas
python-dotenv
pydantic
pydantic-settings
pytest
pytest-cov
pytest-mock
moto
black
flake8
mypy
pylint
cryptography
python-jose
jsonschema
tenacity
kubernetes
pyyaml
aiohttp
